+++ ext/deltalake/src/lib.rs
@@ -11,7 +11,7 @@ use chrono::Duration;
 use deltalake::arrow::ffi_stream::{ArrowArrayStreamReader, FFI_ArrowArrayStream};
 use deltalake::kernel::StructType;
 use deltalake::operations::delete::DeleteBuilder;
-use deltalake::operations::optimize::OptimizeBuilder;
+use deltalake::operations::optimize::{OptimizeBuilder, OptimizeType};
 use deltalake::operations::vacuum::VacuumBuilder;
 use deltalake::storage::IORuntime;
 use deltalake::DeltaOps;
@@ -213,7 +213,6 @@ impl RawDeltaTable {
         Ok(metrics.files_deleted)
     }
 
-    #[allow(clippy::too_many_arguments)]
     pub fn compact_optimize(
         &self,
         target_size: Option<i64>,
@@ -229,6 +228,37 @@ impl RawDeltaTable {
                 .clone(),
         )
         .with_max_concurrent_tasks(max_concurrent_tasks.unwrap_or_else(num_cpus::get));
+        if let Some(size) = target_size {
+            cmd = cmd.with_target_size(size);
+        }
+        if let Some(commit_interval) = min_commit_interval {
+            cmd = cmd.with_min_commit_interval(time::Duration::from_secs(commit_interval));
+        }
+
+        let (table, metrics) = rt().block_on(cmd.into_future()).map_err(RubyError::from)?;
+        self._table.borrow_mut().state = table.state;
+        Ok(serde_json::to_string(&metrics).unwrap())
+    }
+
+    pub fn z_order_optimize(
+        &self,
+        z_order_columns: Vec<String>,
+        target_size: Option<i64>,
+        max_concurrent_tasks: Option<usize>,
+        max_spill_size: usize,
+        min_commit_interval: Option<u64>,
+    ) -> RbResult<String> {
+        let mut cmd = OptimizeBuilder::new(
+            self._table.borrow().log_store(),
+            self._table
+                .borrow()
+                .snapshot()
+                .map_err(RubyError::from)?
+                .clone(),
+        )
+        .with_max_concurrent_tasks(max_concurrent_tasks.unwrap_or_else(num_cpus::get))
+        .with_max_spill_size(max_spill_size)
+        .with_type(OptimizeType::ZOrder(z_order_columns));
         if let Some(size) = target_size {
             cmd = cmd.with_target_size(size);
         }
@@ -389,6 +419,10 @@ fn init(ruby: &Ruby) -> RbResult<()> {
     class.define_method(
         "compact_optimize",
         method!(RawDeltaTable::compact_optimize, 3),
+    )?;
+    class.define_method(
+        "z_order_optimize",
+        method!(RawDeltaTable::z_order_optimize, 5),
     )?;
     class.define_method(
         "update_incremental",