+++ lib/deltalake.rb
@@ -13,6 +13,7 @@ require_relative "deltalake/field"
 require_relative "deltalake/metadata"
 require_relative "deltalake/schema"
 require_relative "deltalake/table"
+require_relative "deltalake/table_alterer"
 require_relative "deltalake/table_optimizer"
 require_relative "deltalake/version"