+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Added support for small file compaction
+- Added support for Z Ordering
 - Added `protocol` method to `Table`
 - Fixed handling of unsigned integers
 - Fixed error with timestamps