+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added support for constraints
 - Added support for small file compaction
 - Added support for Z Ordering
 - Added `protocol` and `repair` methods to `Table`